Compare and classify alternative spliced segments between two or more transcripts

Compare and classify alternative spliced segments between two or more transcripts

compareSplicedSegment(x, ...)

Arguments

x
Can be a GRangesList object containing exons for each transcripts (Intralist mode)

Can be a GRanges object containing exons for a transcript (Pair-wise mode). If so, at least one GRanges object is to be provided in `...` for comparison

...
In pair-wise mode, argument is one or more GRanges object containing exons for a particular transcript to compare with `x`

Value

GRangesList object containing alternatively-spliced segments for each transcript in comparison.

Author

Fursham Hamid

Examples

## ---------------------------------------------------------------------
## EXAMPLE USING TOY DATASET
## ---------------------------------------------------------------------
require(GenomicRanges)

## Create toy GRanges GTF object
gr1 <- GRanges("chr1", IRanges(start = c(1, 101), width = c(20, 20)), "+", transcript_id = "transcript1", gene_id = "geneA", gene_name = "gene001")
gr2 <- GRanges("chr1", IRanges(start = c(1, 51, 101), width = c(20, 10, 20)), "+", transcript_id = "transcript2", gene_id = "geneA", gene_name = "gene001")

## Pairwise comparison between GRanges object
compareSplicedSegment(gr1, gr2)
#> GRangesList object of length 1:
#> $transcript2
#> GRanges object with 1 range and 5 metadata columns:
#>       seqnames    ranges strand |     gene_id   gene_name transcript_id
#>          <Rle> <IRanges>  <Rle> | <character> <character>   <character>
#>   [1]     chr1     51-60      + |       geneA       geneA   transcript2
#>              type      AStype
#>       <character> <character>
#>   [1]          AS          CE
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 

## Multiple comparisons can be done by providing more GRanges input
gr3 <- GRanges("chr1", IRanges(start = c(1, 91), width = c(20, 30)), "+", transcript_id = "transcript3", gene_id = "geneA", gene_name = "gene001")
compareSplicedSegment(gr1, gr2, gr3)
#> GRangesList object of length 2:
#> $transcript2
#> GRanges object with 1 range and 5 metadata columns:
#>       seqnames    ranges strand |     gene_id   gene_name transcript_id
#>          <Rle> <IRanges>  <Rle> | <character> <character>   <character>
#>   [1]     chr1     51-60      + |       geneA       geneA   transcript2
#>              type      AStype
#>       <character> <character>
#>   [1]          AS          CE
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#> $transcript3
#> GRanges object with 1 range and 5 metadata columns:
#>       seqnames    ranges strand |     gene_id   gene_name transcript_id
#>          <Rle> <IRanges>  <Rle> | <character> <character>   <character>
#>   [1]     chr1    91-100      + |       geneA       geneA   transcript3
#>              type      AStype
#>       <character> <character>
#>   [1]          AS          SA
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 

## GRangesList containing exons per transcript can be given as input
grl <- GRangesList(list(gr1, gr2, gr3))
compareSplicedSegment(grl)
#> GRangesList object of length 2:
#> $transcript2
#> GRanges object with 1 range and 5 metadata columns:
#>       seqnames    ranges strand |     gene_id   gene_name transcript_id
#>          <Rle> <IRanges>  <Rle> | <character> <character>   <character>
#>   [1]     chr1     51-60      + |       geneA       geneA   transcript2
#>              type      AStype
#>       <character> <character>
#>   [1]          AS          CE
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#> $transcript3
#> GRanges object with 1 range and 5 metadata columns:
#>       seqnames    ranges strand |     gene_id   gene_name transcript_id
#>          <Rle> <IRanges>  <Rle> | <character> <character>   <character>
#>   [1]     chr1    91-100      + |       geneA       geneA   transcript3
#>              type      AStype
#>       <character> <character>
#>   [1]          AS          SA
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 

## ---------------------------------------------------------------------
## EXAMPLE USING SAMPLE DATASET
## ---------------------------------------------------------------------
## Using GRangesList as input
compareSplicedSegment(query_exons)
#> GRangesList object of length 4:
#> $transcript1
#> GRanges object with 3 ranges and 5 metadata columns:
#>       seqnames            ranges strand |               gene_id
#>          <Rle>         <IRanges>  <Rle> |           <character>
#>   [1]    chr10 79858952-79859071      + | ENSMUSG00000006498.14
#>   [2]    chr10 79860117-79860194      + | ENSMUSG00000006498.14
#>   [3]    chr10 79862014-79862047      + | ENSMUSG00000006498.14
#>                   gene_name transcript_id        type      AStype
#>                 <character>   <character> <character> <character>
#>   [1] ENSMUSG00000006498.14   transcript1          AS          SA
#>   [2] ENSMUSG00000006498.14   transcript1          AS          CE
#>   [3] ENSMUSG00000006498.14   transcript1          AS          CE
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#> $transcript2
#> GRanges object with 2 ranges and 5 metadata columns:
#>       seqnames            ranges strand |               gene_id
#>          <Rle>         <IRanges>  <Rle> |           <character>
#>   [1]    chr10 79858952-79859071      + | ENSMUSG00000006498.14
#>   [2]    chr10 79862014-79862047      + | ENSMUSG00000006498.14
#>                   gene_name transcript_id        type      AStype
#>                 <character>   <character> <character> <character>
#>   [1] ENSMUSG00000006498.14   transcript2          AS          SA
#>   [2] ENSMUSG00000006498.14   transcript2          AS          CE
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#> $transcript3
#> GRanges object with 1 range and 5 metadata columns:
#>       seqnames            ranges strand |               gene_id
#>          <Rle>         <IRanges>  <Rle> |           <character>
#>   [1]    chr10 79858952-79859071      + | ENSMUSG00000006498.14
#>                   gene_name transcript_id        type      AStype
#>                 <character>   <character> <character> <character>
#>   [1] ENSMUSG00000006498.14   transcript3          AS          SA
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#> $transcript4
#> GRanges object with 1 range and 5 metadata columns:
#>       seqnames            ranges strand |               gene_id
#>          <Rle>         <IRanges>  <Rle> |           <character>
#>   [1]    chr10 79862014-79862047      + | ENSMUSG00000006498.14
#>                   gene_name transcript_id        type      AStype
#>                 <character>   <character> <character> <character>
#>   [1] ENSMUSG00000006498.14   transcript4          AS          CE
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 

## Compare AS between individual GRanges object
compareSplicedSegment(query_exons[[1]], query_exons[[3]])
#> GRangesList object of length 1:
#> $transcript1
#> GRanges object with 2 ranges and 5 metadata columns:
#>       seqnames            ranges strand |               gene_id
#>          <Rle>         <IRanges>  <Rle> |           <character>
#>   [1]    chr10 79860117-79860194      + | ENSMUSG00000006498.14
#>   [2]    chr10 79862014-79862047      + | ENSMUSG00000006498.14
#>                   gene_name transcript_id        type      AStype
#>                 <character>   <character> <character> <character>
#>   [1] ENSMUSG00000006498.14   transcript1          AS          CE
#>   [2] ENSMUSG00000006498.14   transcript1          AS          CE
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#>