Generate hypothetical transcripts by adding/removing exons

Adding or removing alternative segments to existing transcripts allow functional investigation of hypothetical RNAs. The functions below accepts two vectors of the same length: x being a GRangesList containing exon coordinates for each transcript to be modified, and y being a GRanges object containing exon coordinates to add/remove to/from transcripts in x. The operations are done in parallel, i.e. first exon in y will be added/remove from the first transcript in x.

addExonstoTx(x, y, drop.unmodified = FALSE, allow.external.exons = FALSE)

removeExonsfromTx(x, y, drop.unmodified = FALSE, ignore.strand = FALSE)

Arguments

x
GRangesList containing exon coordinates of each transcript to be modified

y
GRanges object containing exon coordinates to be added/removed in parallel to transcripts in x. Length of y and x have to be equal. In addition, x-y pairs have to be on the same chromosome and strand.

drop.unmodified
Whether to drop hypothetical transcripts which were not successfully modified. (Default; FALSE)

allow.external.exons
Whether to allow insertion of exons that fall outside the range of the transcript (Default: FALSE). We recommend carefully inspecting the hypothetical transcripts when switching this argument to TRUE, as some exons may not originate from the same gene.

ignore.strand
Whether to ignore strand when removing segments from transcript (Default: FALSE)

Value

GRangesList containing modified transcripts in x. If drop.unmodified is FALSE, output object is of same length as x and and contain metadata `modified` to annotate whether insertion/removal is successful. All other metadata in x will not be retained

Examples

## ---------------------------------------------------------------------
## EXAMPLE USING TOY DATASET
## ---------------------------------------------------------------------
require(GenomicRanges)

#> Loading required package: GenomicRanges

#> Loading required package: stats4

#> Loading required package: BiocGenerics

#> Loading required package: parallel

#> 
#> Attaching package: ‘BiocGenerics’

#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from ‘package:base’:
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colnames,
#>     dirname, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, Map, mapply, match, mget,
#>     order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rownames, sapply, setdiff, sort, table, tapply,
#>     union, unique, unsplit, which.max, which.min

#> Loading required package: S4Vectors

#> 
#> Attaching package: ‘S4Vectors’

#> The following object is masked from ‘package:base’:
#> 
#>     expand.grid

#> Loading required package: IRanges

#> Loading required package: GenomeInfoDb

# create GRangesList of transcript exon coordinates
gr1 <- GRanges(
seqnames = "1", strand = rep("+", 2),
ranges = IRanges::IRanges(
start = c(1, 501),end = c(100, 600)))

grl1 <- GRangesList(gr1, gr1, gr1)
names(grl1) <- c("tx1", "tx2", "tx3")

# add casette exon, alternative 3'ss segment and retained intron
exons <- GenomicRanges::GRanges(
seqnames = "1", strand = rep("+", 3),
ranges = IRanges::IRanges(start = c(201, 401, 101),
                          end = c(300, 500, 500)))
                          
mod.grl1 <- addExonstoTx(grl1, exons)
 
# remove exons from hypothetical transcripts
removeExonsfromTx(mod.grl1, exons) 

#> GRangesList object of length 3:
#> $tx1
#> GRanges object with 2 ranges and 1 metadata column:
#>       seqnames    ranges strand |  modified
#>          <Rle> <IRanges>  <Rle> | <logical>
#>   [1]        1     1-100      + |      TRUE
#>   [2]        1   501-600      + |      TRUE
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#> $tx2
#> GRanges object with 2 ranges and 1 metadata column:
#>       seqnames    ranges strand |  modified
#>          <Rle> <IRanges>  <Rle> | <logical>
#>   [1]        1     1-100      + |      TRUE
#>   [2]        1   501-600      + |      TRUE
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#> $tx3
#> GRanges object with 2 ranges and 1 metadata column:
#>       seqnames    ranges strand |  modified
#>          <Rle> <IRanges>  <Rle> | <logical>
#>   [1]        1     1-100      + |      TRUE
#>   [2]        1   501-600      + |      TRUE
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#>