Find NMD-causing exons from GTF objects

This function will identify alternative exons that triggers NMD when spliced or skipped.

identifyNMDexons(x, fasta, NMD.result = NULL)

Arguments

x

GRanges object containing query GTF data with CDS information.

fasta

BSgenome or Biostrings object containing genomic sequence

NMD.result

Output dataframe from predictNMD() [Optional]. If not provided, or if data-frame do not contain necessary variables, identifyNMDexons will run predictNMD() first.

Value

GRanges object containing coordinates of identified NMD-causing exons, including metadata on its gene of origin, type of AS exon, type of NMD exon and whether it is spliced within CDS

Details

Alternative splicing of regulated exons may introduce premature termination codon which are cleared by nonsense-mediated decay. There are two major class of such exons: Poison and ORF-maintaining exons. The former triggers NMD when spliced in, hence its name, while the latter triggers NMD when removed from a transcript.

To identify and classify NMD-causing exons, identifyNMDexons first find alternative segments between NMD-sensitive transcripts and pre-determined reference transcript of the same gene. The longest, NMD-insensitive coding transcript of each gene is chosen as the best reference. For each alternative segment, a modified version of reference RNA is built by inserting exons spliced in NMD transcripts, or by removing exons skipped from NMD transcripts. These modified transcripts are then tested for NMD features and if positive, the alternative exon will be annotated as NMD-causing.

Examples

## ---------------------------------------------------------------------
## EXAMPLE USING SAMPLE DATASET
## ---------------------------------------------------------------------
library(BSgenome.Mmusculus.UCSC.mm10)

# build CDS information first
out.gtf <- buildCDS(matched_query_gtf, ref_gtf, Mmusculus)
#> Searching for reference mRNAs in query
#> 2 reference mRNAs found and its CDS were assigned
#> Building database of annotated ATG codons
#> Selecting best ATG start codon for remaining transcripts and determining open-reading frame
#> 2 new CDSs constructed
#> 
#> Summary: Out of 4 transcripts in `matched_query_gtf`, 4 transcript CDSs were built

# proceed with identifying NMD exons
NMD.exons_1 <- identifyNMDexons(out.gtf, Mmusculus)
#> `NMD.result` not provided, running predictNMD()
#> Selecting best reference mRNAs
#> Error in dplyr::slice_max(., order_by = ".", n = 1): Problem while computing indices.
#> ℹ The error occurred in group 1: gene_id = "ENSMUSG00000006498.14".
#> Caused by error:
#> ! `order_by` must have size 83, not size 1.

# if output of predictNMD() has been generated, this can be provided to
# NMD.result argument
NMD.output <- predictNMD(out.gtf)
#> Predicting NMD sensitivities for 4 mRNAs
NMD.exons_2 <- identifyNMDexons(out.gtf, Mmusculus, NMD.result = NMD.output)
#> Selecting best reference mRNAs
#> Error in dplyr::slice_max(., order_by = ".", n = 1): Problem while computing indices.
#> ℹ The error occurred in group 1: gene_id = "ENSMUSG00000006498.14".
#> Caused by error:
#> ! `order_by` must have size 83, not size 1.

identical(NMD.exons_1, NMD.exons_2)
#> Error in identical(NMD.exons_1, NMD.exons_2): object 'NMD.exons_1' not found

## ---------------------------------------------------------------------
## EXAMPLE USING TRANSCRIPT ANNOTATION
## ---------------------------------------------------------------------
# \donttest{
library(AnnotationHub)
#> Loading required package: BiocFileCache
#> Loading required package: dbplyr
#> 
#> Attaching package: ‘AnnotationHub’
#> The following object is masked from ‘package:rtracklayer’:
#> 
#>     hubUrl

## Retrieve GRCm38 trancript annotation
ah <- AnnotationHub()
#> snapshotDate(): 2021-10-20
GRCm38_gtf <- ah[["AH60127"]]
#> downloading 1 resources
#> retrieving 1 resource
#> loading from cache
#> Importing File into R ..
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> '?repositories' for details
#> 
#> replacement repositories:
#>     CRAN: https://cloud.r-project.org
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> '?repositories' for details
#> 
#> replacement repositories:
#>     CRAN: https://cloud.r-project.org

## Identify NMD-causing exons in entire transcriptome
NMD.exons.all <- identifyNMDexons(GRCm38_gtf, Mmusculus)
#> Error in .identifynmdchecks(x, fasta, NMD.result, argnames): `GRCm38_gtf` and `Mmusculus` has unmatched seqlevel styles. 
#> Try running: GRCm38_gtf <- matchChromosomes(GRCm38_gtf, Mmusculus)
# }